Implemented many features of the GrindableObject prefab
Prefab is able to identify the player, and tell them to snap to the prefab's x position and enter a grinding state while in contact with its top trigger box. While in this state, the player cannot move left or right, but jumping or going off of the far end of the object takes them out of the state.

The object also moves toward the player, and teleports back in front of them after passing by. This is mostly a test thing.