Added diving after a jump
Ignore commented-out code in PhysicsObject, as everything for this is in FishMovement. This also had the neat side effect of letting me clean up some variables in favor of an enum.

I also disabled the collider on road slices, and added a temporary transparent blue texture to them so you can see the player through them.