Added perfect dismount to grind rails, disabled jumping while underwater (see desc) and cleaned up some old Debug.Log calls
I kind of like being able to "jump" underwater in order to quickly return to the surface, but it seemed to be causing an issue where jumping at exactly the right time when surfacing caused the player to shoot up super high into the air (funny, but probably needed to be fixed). Will revisit.